Improvements to NewtonSolve, plus corresponding tests.
Applied more default values so that the maximum acceptable error and maximum
number of iterations are both configurable.  And then changed the behaviour
when the maximum number of iterations is exceeded to throw an exception rather
than the previous behaviour of returning a nonsense number.  And this part got
some tests to check that it works properly.

Okay, I suppose that should have been two commits, but come on, it's not that
big.